requirements.txt
setuptools==41.0.0
numpy==1.13.3
pandas==0.23.4
Keras==2.2.4
scipy==1.1.0
python-dotenv==0.10.3
scikit_learn==0.21.2
PyYAML==3.13
